refactor: remove confusing collapse component
